admin: improve logging and remove rp_service.h
- Log additional context when there is an issue restarting the registry
- Log the service parameter instead of service_kind
- Explictly call to_string_view when logging service_kind
- Move the contents of rp_service.h into admin_server.h

(cherry picked from commit 4b78face437cc87ce929540adcd87b8d25949440)
