+++ src/v/redpanda/admin_server.h
@@ -17,7 +17,6 @@
 #include "model/metadata.h"
 #include "pandaproxy/schema_registry/fwd.h"
 #include "request_auth.h"
-#include "rp_services.h"
 #include "rpc/connection_cache.h"
 #include "seastarx.h"
 
@@ -37,6 +36,10 @@ struct admin_server_cfg {
     std::vector<config::endpoint_tls_config> endpoints_tls;
     ss::sstring admin_api_docs_dir;
     ss::scheduling_group sg;
+};
+
+enum class service_kind {
+    schema_registry,
 };
 
 namespace detail {
